demo: use register_custom_background_input_handler() to support keyboard shortcuts
THIS IS A DEMO: IT IS NOT INTENDED TO BE COMMITTED

This provides an example of how `register_custom_background_input_handler()`
could be used to add keyboard shortcuts for things like switching verbs.
As you can see, this adds a number of entries to the input map in
`project.godot` to add keyboard shortcuts. If you're curious, I chose
the current set of shortcuts to match the geometrical layout in the
9verb UI:

```
open  | pickup | push    ->    Q | W | E
close | look   | pull    ->    A | S | D
give  | use    | talk    ->    Z | X | C
```

The chosen letters are not great mnemonic devices, but it does make it
easier to play left-hand-keyboard/right-hand-mouse (Thimbleweed Park
supports this, fwiw.)

From this experiment, it seems to me that we need something like
`register_custom_background_input_handler()` in escoria-core to
facilitate building custom UI like this.