feat: ESCItem can specify Area2D of get_interact_position()
The goal of this change is to make the player's proximity to an
`ESCItem` "more forgiving" in order to act on it.

At a high level, this commit introduces a new method on `ESCItem`,
`can_interact_from()`, that is used by
`ESCActionManager.perform_inputevent_on_object()` to determine whether
the player is "close enough" to perform an action.
It also defines a new, optional property on `ESCItem` named
`interact_node`, which is a `NodePath` to an `Area2D` that is used by
`can_interact_from()`.

Here is some of the logic behind the change to `esc_action_manager.gd`
to verify that existing behavior is preserved. Specifically, this line
appears safe to remove because `clicked_position` was not read:

```gdscript
var clicked_position = event.position
```

Removing that line left us with:

```gdscript
		var destination_position: Vector2 = escoria.main.current_scene.player.\
				global_position

    # BIG CONDITIONAL

		var player_global_pos = escoria.main.current_scene.player.global_position

		if not player_global_pos.is_equal_approx(destination_position):
			dont_interact = true
```

In the case where `BIG CONDITIONAL` is `false`, notice that
`destination_position` and `player_global_pos` are the result of the
same expression, in which case this condition will never be `true`:

```gdscript
		if not player_global_pos.is_equal_approx(destination_position):
			dont_interact = true
```

This means that it is safe to make this check part of `BIG CONDITIONAL`
because `player_global_pos` is not used later in the function, either.

It follows that the logic that could set `dont_interact` applies only in
the specific circumstance where we started walking the player towards
an object, but we have to make sure they made it there (e.g., the
user could have redirected the player in the meantime) and are
"close enough" to act on the `ESCObject`.

We replace the existing `is_equal_approx()` check with a call to the
new `can_interact_from()` method, which is overridable by an individual
`ESCItem`. Though we hope to make that unnecessary, in practice, by
giving the caller the ability to specify an `Area2D` as the
`interact_node` property to define the valid area where the player can
be standing in order to interact with an `ESCItem`.