fix: move ESCCondition parsing logic out of the constructor
This makes it possible to create an `ESCCondition` programmatically.
Unfortunately, this does not fix the issue with `ESCCondition`
where `comparison_regex.compile(REGEX)` is called *every time* an
`ESCCondition` is instantiated.

I have read all sorts of forums, and no one seems to address this
issue head-on. The question of, "Why doesn't GDScript support static
variables" is not addressed clearly in the docs, IMHO.

The only "workaround" I can find is this article on Singletons/Autoload:

https://docs.godotengine.org/en/latest/tutorials/scripting/singletons_autoload.html

Though this seems incredibly heavyweight for caching a `RegEx`.